// Frontend/src/lib/types.ts
export type Role = "READER" | "ADMIN";

export interface User {
  id: string;
  name: string;
  email: string;
  role: Role;
  createdAt: string;
}

export interface Book {
  id: string;
  title: string;
  author: string;
  category: string;
  cover: string;
  description: string;
  year: number;
  pages: string[]; // simulated pages of content
  fileUrl: string; // download
  fileExt?: string; // pdf, epub, txt etc
}

export interface Review {
  id: string;
  bookId: string;
  userId: string;
  userName: string;
  rating: number;
  comment: string;
  createdAt: string;
}